Automate and fix assembly creation.
Notable changes:
* Poms are no longer in git, since they're generated by `./bin/fiji-build generate-pom`
* `./bin/fiji-build build` implies `./bin/fiji-build generate-pom`
* Deployment details are now included in the generated pom's template
* The final version of fiji is now extracted to `FIJI_VERSION`
* A compile fix needed for assembly, which is condicted directly by `mvn`.
* `devtools/create-assemblies-and-deploy-jars` creates assemblies (tarballs).
* Bump to version `0.1.3`
* Travis deployment clauses deploy to github releases and maven central

The `create-assemblies-and-deploy-jars` script uses `./bin/fiji-maven` wrapper around `mvn`.

Assembly descriptors have been altered to include build artifacts in tarballs by copying them directly from the local maven repo.  This is preferred to calling `mvn package`, as the package-made jars don't actually have the same content sometimes.

+++ fiji-schema/fiji-schema/src/test/java/com/moz/fiji/schema/TestFinalColumns.java
@@ -71,10 +71,10 @@ public class TestFinalColumns extends FijiClientTest {
         final FijiTableReader reader = table.openTableReader();
         try {
           final FijiRowData row = reader.get(eid, FijiDataRequest.create("family"));
-          assertEquals(314159, row.getMostRecentValue("family", "int"));
+          assertEquals(314159, (int) row.getMostRecentValue("family", "int"));
 
           // Ensures the long value has not been overwritten:
-          assertEquals(314159L, row.getMostRecentValue("family", "long"));
+          assertEquals(314159L, (long) row.getMostRecentValue("family", "long"));
         } finally {
           reader.close();
         }
